refactor(utils): extract money/json/zod into shared @agicash/utils package
Move the framework-free money/, json, and zod utilities out of
apps/web-wallet/app/lib into a new @agicash/utils leaf package so the upcoming
@agicash/wallet-sdk can depend on them without importing from the web app.

- packages/utils: new @agicash/-scoped, source-linked package (money, json, zod).
  exports use a "./*": "./src/*.ts" wildcard plus an explicit "./money" entry
  (the wildcard cannot index-resolve the money/ directory under Bundler
  resolution; the bare "." root also needs its own entry)
- money/json/zod moved byte-identical via git mv; logic untouched (money is
  financial code)
- old ~/lib/{money,json,zod} paths kept as transitional re-export shims so the
  ~80 consumers stay untouched; a later cleanup PR rewrites the imports and
  deletes the shims
- big.js, @types/big.js, zod hoisted into the root workspaces.catalog (now
  shared by >=2 packages) and referenced via catalog:
- web-wallet now depends on @agicash/utils (workspace:*)

Phase 0 of the wallet-sdk extraction. typecheck, biome, tests (utils 14, web 119)
and full build all green; zero web behavior change.
