+++ apps/web-wallet/app/lib/money/index.ts
@@ -1,2 +1,2 @@
-export { Money } from './money';
-export type { Currency, CurrencyUnit } from './types';
+// Transitional re-export — moved to @agicash/utils; removed in the import-cleanup PR.
+export * from '@agicash/utils/money';
