+++ apps/web-wallet/package.json
@@ -21,6 +21,7 @@
     "@agicash/breez-sdk-spark": "0.13.5-1",
     "@agicash/opensecret": "catalog:",
     "@agicash/qr-scanner": "0.1.2",
+    "@agicash/utils": "workspace:*",
     "@cashu/cashu-ts": "3.6.1",
     "@noble/ciphers": "1.3.0",
     "@noble/curves": "1.9.7",
@@ -49,7 +50,7 @@
     "@tanstack/react-query-devtools": "5.91.3",
     "@vercel/analytics": "1.5.0",
     "@vercel/react-router": "1.3.0",
-    "big.js": "7.0.1",
+    "big.js": "catalog:",
     "class-variance-authority": "0.7.1",
     "clsx": "2.1.1",
     "deepmerge": "4.3.1",
@@ -69,7 +70,7 @@
     "tailwindcss-animate": "1.0.7",
     "usehooks-ts": "3.1.1",
     "vaul": "1.1.2",
-    "zod": "4.3.6",
+    "zod": "catalog:",
     "zustand": "5.0.11"
   },
   "devDependencies": {
@@ -77,7 +78,7 @@
     "@stablelib/base64": "catalog:",
     "@stablelib/chacha20poly1305": "catalog:",
     "@tailwindcss/vite": "4.1.18",
-    "@types/big.js": "6.2.2",
+    "@types/big.js": "catalog:",
     "@types/bun": "catalog:",
     "@types/express": "5.0.0",
     "@types/jwt-encode": "1.0.3",
