+++ packages/utils/src/money/index.ts
@@ -1,0 +1,2 @@
+export { Money } from './money';
+export type { Currency, CurrencyUnit } from './types';
